refactor: merge *_security_test.go into their respective *_test.go files - moved helpers below tests
